requirements.txt
customtkinter
tkinterdnd2
pymupdf
reportlab
requests
Pillow
pymupdf
